+++ R/eppasm.R
@@ -30,10 +30,18 @@ simmod.specfp <- function(fp, VERSION="C"){
   pregprevlag <- rep(0, PROJ_YEARS)
 
   ## initialize projection
+<<<<<<< HEAD
   pop <- array(0, c(pAG, NG, pDS, PROJ_YEARS))                            ### By age group, gender, disease stage (+/-), years
   pop[,,1,1] <- fp$basepop                                                ### Put in the males and females of each age group during start year
   hivpop <- array(0, c(hDS, hAG, NG, PROJ_YEARS))                         ### 7 dis stage, 9 ages, 2 genders, 52 years
   artpop <- array(0, c(hTS, hDS, hAG, NG, PROJ_YEARS))                    ### 3 treatment step, 7 dis stage, 9 age, 2 gen, 52 years
+=======
+  pop <- array(0, c(pAG, NG, pDS, PROJ_YEARS))
+  pop[,,1,1] <- fp$basepop
+  hivpop <- array(0, c(hDS, hAG, NG, PROJ_YEARS))
+  diagnpop <- array(0, c(hDS, hAG, NG, PROJ_YEARS))
+  artpop <- array(0, c(hTS, hDS, hAG, NG, PROJ_YEARS))
+>>>>>>> 0a8320e8c0569fb4bd1e74ba7689baaee56b19ff
 
   ## initialize output
   prev15to49 <- numeric(PROJ_YEARS)                                       ### Initialize 52 0s for prevalence from 15 to 49 over time period
@@ -45,7 +53,15 @@ simmod.specfp <- function(fp, VERSION="C"){
   hivdeaths <- array(0, c(pAG, NG, PROJ_YEARS))                           ### array of hiv deaths by age, gender, year of epi
   natdeaths <- array(0, c(pAG, NG, PROJ_YEARS))                           ### array of natural deaths in pop
 
+<<<<<<< HEAD
   popadj.prob <- array(0, c(pAG, NG, PROJ_YEARS))                         ### array for adjust by age group, gender, year of epi
+=======
+  diagnoses <- array(0, c(hDS, hAG, NG, PROJ_YEARS))
+  artinits <- array(0, c(hDS, hAG, NG, PROJ_YEARS))
+  
+
+  popadj.prob <- array(0, c(pAG, NG, PROJ_YEARS))
+>>>>>>> 0a8320e8c0569fb4bd1e74ba7689baaee56b19ff
 
   entrant_prev_out <- numeric(PROJ_YEARS)                                 ### prevalence for each year of epi
   hivp_entrants_out <- array(0, c(NG, PROJ_YEARS))                        ### number of hiv pos by gender and year of epi
@@ -104,6 +120,12 @@ simmod.specfp <- function(fp, VERSION="C"){
     hivpop[,-1,,i] <- hivpop[,-1,,i] + sweep(hivpop[,-hAG,,i-1], 2:3, hiv.ag.prob[-hAG,], "*")   ## This adds in survival prob to next group
     hivpop[,1,,i] <- hivpop[,1,,i] + sweep(fp$paedsurv_cd4dist[,,i], 2, hivp_entrants * (1-fp$entrantartcov[,i]), "*")
 
+    if(i > fp$t_diagn_start){
+      diagnpop[,,,i] <- diagnpop[,,,i-1]
+      diagnpop[,-hAG,,i] <- diagnpop[,-hAG,,i] - sweep(diagnpop[,-hAG,,i-1], 2:3, hiv.ag.prob[-hAG,], "*")
+      diagnpop[,-1,,i] <- diagnpop[,-1,,i] + sweep(diagnpop[,-hAG,,i-1], 2:3, hiv.ag.prob[-hAG,], "*")
+    }
+    
     if(i > fp$tARTstart){
       artpop[,,,,i] <- artpop[,,,,i-1]
       artpop[,,-hAG,,i] <- artpop[,,-hAG,,i] - sweep(artpop[,,-hAG,,i-1], 3:4, hiv.ag.prob[-hAG,], "*") ## Age everyone out 
@@ -118,8 +140,14 @@ simmod.specfp <- function(fp, VERSION="C"){
     pop[,,,i] <- pop[,,,i] - deaths                                                    ## Initially works out death matrix then -s
     natdeaths[,,i] <- rowSums(deaths,,2)
 
+<<<<<<< HEAD
     hivpop[,,,i] <- sweep(hivpop[,,,i], 2:3, hiv.sx.prob, "*")                         ## Works out next hiv pop based on natural
                                                                                        ## deaths
+=======
+    hivpop[,,,i] <- sweep(hivpop[,,,i], 2:3, hiv.sx.prob, "*")
+    if(i > fp$t_diagn_start)
+      diagnpop[,,,i] <- sweep(diagnpop[,,,i], 2:3, hiv.sx.prob, "*")
+>>>>>>> 0a8320e8c0569fb4bd1e74ba7689baaee56b19ff
     if(i > fp$tARTstart)
       artpop[,,,,i] <- sweep(artpop[,,,,i], 3:4, hiv.sx.prob, "*")
 
@@ -131,6 +159,8 @@ simmod.specfp <- function(fp, VERSION="C"){
     pop[,,,i] <- sweep(pop[,,,i], 1:2, mr.prob, "*")                                   ## Assumes migration matches NL HIV status ratio
 
     hivpop[,,,i] <- sweep(hivpop[,,,i], 2:3, hiv.mr.prob, "*")
+    if(i > fp$t_diagn_start)
+      diagnpop[,,,i] <- sweep(diagnpop[,,,i], 2:3, hiv.mr.prob, "*")
     if(i > fp$tARTstart)
       artpop[,,,,i] <- sweep(artpop[,,,,i], 3:4, hiv.mr.prob, "*")                    ## Assumes same migration rate as for HIV +ve
 
@@ -162,7 +192,7 @@ simmod.specfp <- function(fp, VERSION="C"){
           rvec[ts] <- calc_rtrend_rt(fp$proj.steps[ts], fp, rvec[ts-1], prevlast, pop, i, ii)
         else
           rvec[ts] <- fp$rvec[ts]
-
+        
         ## number of infections by age / sex
         if(exists("incidmod", where=fp) && fp$incidmod == "transm")
           infections.ts <- calc_infections_simpletransm(fp, pop, hivpop, artpop, i, ii, rvec[ts])
@@ -182,6 +212,7 @@ simmod.specfp <- function(fp, VERSION="C"){
       }
 
       ## disease progression and mortality
+<<<<<<< HEAD
       grad[-hDS,,] <- grad[-hDS,,] - fp$cd4_prog * hivpop[-hDS,,,i]  # remove cd4 stage progression (untreated) all can be removed but last class
       grad[-1,,] <- grad[-1,,] + fp$cd4_prog * hivpop[-hDS,,,i]      # add cd4 stage progression (untreated) all can be jumped into but last class
 
@@ -193,10 +224,35 @@ simmod.specfp <- function(fp, VERSION="C"){
       hivdeaths_p.ts <- apply(hivdeaths.ts, 2, rep, h.ag.span) * apply(pop[,,hivp.idx,i], 2, calc.agdist)  # HIV deaths by single-year age
       pop[,,2,i] <- pop[,,2,i] - hivdeaths_p.ts
       hivdeaths[,,i] <- hivdeaths[,,i] + hivdeaths_p.ts
+=======
+      grad[-hDS,,] <- grad[-hDS,,] - fp$cd4_prog * hivpop[-hDS,,,i]  # remove cd4 stage progression (untreated)
+      grad[-1,,] <- grad[-1,,] + fp$cd4_prog * hivpop[-hDS,,,i]      # add cd4 stage progression (untreated)
+      
+      grad <- grad - fp$cd4_mort * hivpop[,,,i]              # HIV mortality, untreated
+      hivdeaths.ts <- colSums(fp$cd4_mort * hivpop[,,,i])
+>>>>>>> 0a8320e8c0569fb4bd1e74ba7689baaee56b19ff
 
       hivpop[,,,i] <- hivpop[,,,i] + DT*grad
 
-      ## ART initiation
+      ## diagnosed population: diagnosis, progression, and mortality
+      if(i >= fp$t_diagn_start){
+
+        grad_diagn <- array(0, c(hDS, hAG, NG))
+
+        new_diagn <- fp$diagn_rate[,,,i] * (hivpop[,,,i] - diagnpop[,,,i])
+        grad_diagn <- grad_diagn + new_diagn
+
+        grad_diagn[-hDS,,] <- grad_diagn[-hDS,,] - fp$cd4_prog * diagnpop[-hDS,,,i]  # remove cd4 stage progression (untreated)
+        grad_diagn[-1,,] <- grad_diagn[-1,,] + fp$cd4_prog * diagnpop[-hDS,,,i]      # add cd4 stage progression (untreated)
+        
+        grad_diagn <- grad_diagn - fp$cd4_mort * diagnpop[,,,i]                      # HIV mortality, untreated
+
+        diagnpop[,,,i] <- diagnpop[,,,i] + DT * grad_diagn
+        diagnoses[,,,i] <- diagnoses[,,,i] + DT * new_diagn
+      }
+
+        
+      ## ART population
       if(i >= fp$tARTstart) {
 
         gradART <- array(0, c(hTS, hDS, hAG, NG))
@@ -207,14 +263,22 @@ simmod.specfp <- function(fp, VERSION="C"){
 
         gradART <- gradART - fp$art_mort * artpop[,,,,i]                  # ART mortality have a table for mort, doesn't change through time
 
-        ## hivdeaths.ts <- hivdeaths.ts + colSums(fp$art_mort * artpop[,,,,i],,2)
+        hivdeaths.ts <- hivdeaths.ts + colSums(fp$art_mort * artpop[,,,,i],,2)
         artpop[,,,, i] <- artpop[,,,, i] + DT * gradART
 
 
         ## ART dropout
         ## remove proportion from all adult ART groups back to untreated pop
+<<<<<<< HEAD
         hivpop[,,,i] <- hivpop[,,,i] + DT*fp$art_dropout[i]*colSums(artpop[,,,,i])     ## Means treats ART pop as separate, not nested!!
         artpop[,,,,i] <- artpop[,,,,i] - DT*fp$art_dropout[i]*artpop[,,,,i]            ## ART pop isn't nested within HIV pop
+=======
+        hivpop[,,,i] <- hivpop[,,,i] + DT*fp$art_dropout[i]*colSums(artpop[,,,,i])
+        if(i >= fp$t_diagn_start)
+          diagnpop[,,,i] <- diagnpop[,,,i] + DT*fp$art_dropout[i]*colSums(artpop[,,,,i])
+
+        artpop[,,,,i] <- artpop[,,,,i] - DT*fp$art_dropout[i]*artpop[,,,,i]
+>>>>>>> 0a8320e8c0569fb4bd1e74ba7689baaee56b19ff
 
         ## calculate number eligible for ART
         artcd4_percelig <- 1 - (1-rep(0:1, times=c(fp$artcd4elig_idx[i]-1, hDS - fp$artcd4elig_idx[i]+1))) *
@@ -306,9 +370,24 @@ simmod.specfp <- function(fp, VERSION="C"){
 
         hivpop[, h.age15plus.idx,, i] <- hivpop[, h.age15plus.idx,, i] - artinit
         artpop[1,, h.age15plus.idx,, i] <- artpop[1,, h.age15plus.idx,, i] + artinit
-      }
-    }
+        
+        if(i >= fp$t_diagn_start){
+          newdiagn <- pmax(artinit - diagnpop[,,,i], 0)
+          diagnoses[,,,i] <- diagnoses[,,,i] + newdiagn
+          diagnpop[,,,i] <- diagnpop[,,,i] - (artinit - newdiagn)
+        }
 
+        artinits[,,,i] <- artinits[,,,i] + artinit
+      }
+        
+      ## Remove hivdeaths from pop
+      calc.agdist <- function(x) {d <- x/rep(ctapply(x, ag.idx, sum), h.ag.span); d[is.na(d)] <- 0; d}
+      hivdeaths_p.ts <- apply(DT*hivdeaths.ts, 2, rep, h.ag.span) * apply(pop[,,hivp.idx,i], 2, calc.agdist)  # HIV deaths by single-year age
+      pop[,,2,i] <- pop[,,2,i] - hivdeaths_p.ts
+      hivdeaths[,,i] <- hivdeaths[,,i] + hivdeaths_p.ts
+
+    }
+      
     ## ## Code for calculating new infections once per year to match prevalence (like Spectrum)
     ## ## incidence
     ## prev.i <- sum(pop[p.age15to49.idx,,2,i]) / sum(pop[p.age15to49.idx,,,i]) # prevalence age 15 to 49
@@ -340,7 +419,9 @@ simmod.specfp <- function(fp, VERSION="C"){
 
       pop[,,,i] <- sweep(pop[,,,i], 1:2, popadj.prob[,,i], "*")
       hivpop[,,,i] <- sweep(hivpop[,,,i], 2:3, hiv.popadj.prob, "*")
-      if(t >= fp$tARTstart)
+      if(i >= fp$t_diagn_start)
+        diagnpop[,,,i] <- sweep(diagnpop[,,,i], 2:3, hiv.popadj.prob, "*")
+      if(i >= fp$tARTstart)
         artpop[,,,,i] <- sweep(artpop[,,,,i], 3:4, hiv.popadj.prob, "*")
 
     }
@@ -363,11 +444,15 @@ simmod.specfp <- function(fp, VERSION="C"){
   attr(pop, "incid15to49") <- incid15to49
   attr(pop, "sexinc") <- sexinc15to49out
   attr(pop, "hivpop") <- hivpop
+  attr(pop, "diagnpop") <- diagnpop
   attr(pop, "artpop") <- artpop
 
   attr(pop, "infections") <- infections
   attr(pop, "hivdeaths") <- hivdeaths
   attr(pop, "natdeaths") <- natdeaths
+
+  attr(pop, "diagnoses") <- diagnoses
+  attr(pop, "artinits") <- artinits
 
   attr(pop, "popadjust") <- popadj.prob