+++ NEWS.md
@@ -92,3 +92,8 @@
 # eppasm 0.5.1
 
 - Add tidy_output() function to generate long format of key output indicators.
+
+# eppasm 0.6.0
+
+- Explicitly model new diagnoses.
+- Add number of new diagnoses and ART initiations by age, sex, and CD4 category as model outputs.